# hospital-admin-back-swagger.yml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: https://localhost:8080
    description: Generated server url
paths:
  /api/v1/admin/patients:
    get:
      tags:
        - patient-controller
      summary: Get list of patients
      operationId: getPatients
      parameters:
        - name: name
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Get list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListPatientDTO'
        '400':
          description: Incorrect values (SQL injection, for example). Full description in "message" field
          content: {}
    post:
      tags:
        - patient-controller
      summary: Create new patient
      operationId: addPatient
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientLongDTO'
        required: true
      responses:
        '200':
          description: Patient created, returning patient with id field
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientLongDTO'
        '400':
          description: Incorrect values (SQL injection, for example). Full description in "message" field
          content: {}
  /api/v1/admin/patient/{id}:
    get:
      tags:
        - patient-controller
      summary: Get specific patient
      operationId: getPatient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Get patient
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientLongDTO'
        '400':
          description: No patient with such id. Full description in "message" field
          content: {}
components:
  schemas:
    PatientLongDTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        surname:
          type: string
        registrationTime:
          type: string
          format: date-time
        socialStatus:
          type: string
        isMale:
          type: boolean
        isMage:
          type: boolean
    ListPatientDTO:
      type: object
      properties:
        patients:
          type: array
          items:
            $ref: '#/components/schemas/PatientShortDTO'
    PatientShortDTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        surname:
          type: string